+++ packages/graph-explorer/src/core/StateProvider/useResetState.ts
@@ -21,6 +21,8 @@ import {
 import { isRestorePreviousSessionAvailableAtom } from "./graphSession";
 import { RESET, useAtomCallback } from "jotai/utils";
 import { useCallback } from "react";
+import { selectedTabAtom } from "@/modules/SearchSidebar";
+import { queryTextAtom } from "@/modules/SearchSidebar/QuerySearchTabContent";
 
 export default function useResetState() {
   return useAtomCallback(
@@ -44,6 +46,10 @@ export default function useResetState() {
       set(selectedVertexTypeAtom, RESET);
       set(selectedAttributeAtom, RESET);
       set(partialMatchAtom, RESET);
+
+      // Query editor
+      set(selectedTabAtom, RESET);
+      set(queryTextAtom, RESET);
 
       // Previous session
       set(isRestorePreviousSessionAvailableAtom, RESET);